Fix: request 함수와 api 분리
